pyproject.toml
[tool.poetry]
name = "travel-monitor"
version = "0.1.0"
description = "A smart travel price monitor that tracks flight and hotel packages to your desired destinations and sends email alerts when the best deals are available."
authors = ["Felipe Merenda <felipemerenda1@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
amadeus = "^9.0.0"
pyyaml = "^6.0"
python-dotenv = "^1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0"
mypy = "^1.0"
ruff = "^0.4"
types-pyyaml = "^6.0.12.20250915"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
